+++ include/mpi/gen_collective_comm.sh
@@ -7,7 +7,7 @@ for prec in double float int size_t; do
   * @brief MPI_Allreduce (MPI_SUM) for scalar. Combines values from all processes and distributes the result back to all processes.
   * @param val scalar value
   */ "
-  echo "[[nodiscard]] $prec Allreduce ($prec val) const;"
+  echo "[[nodiscard]] $prec Allreduce($prec val) const;"
 done
 
 for prec in double float int size_t; do
@@ -16,7 +16,7 @@ for prec in double float int size_t; do
   * @brief MPI_Allreduce (MPI_SUM) for scalar. Combines values from all processes and distributes the result back to all processes.
   * @param val scalar value
   */ "
-  echo "[[nodiscard]] $prec Allreduce_sum ($prec val) const;"
+  echo "[[nodiscard]] $prec Allreduce_sum($prec val) const;"
 done
 
 for prec in double float int size_t; do
@@ -25,7 +25,7 @@ for prec in double float int size_t; do
   * @brief MPI_Allreduce (MPI_PROD) for scalar. Combines values from all processes and distributes the result back to all processes.
   * @param val scalar value
   */ "
-  echo "[[nodiscard]] $prec Allreduce_prod ($prec val) const;"
+  echo "[[nodiscard]] $prec Allreduce_prod($prec val) const;"
 done
 
 for prec in double float int size_t; do
@@ -34,7 +34,7 @@ for prec in double float int size_t; do
   * @brief MPI_Allreduce (MPI_MAX) for scalar. Combines values from all processes and distributes the result back to all processes.
   * @param val scalar value
   */ "
-  echo "[[nodiscard]] $prec Allreduce_max ($prec val) const;"
+  echo "[[nodiscard]] $prec Allreduce_max($prec val) const;"
 done
 
 for prec in double float int size_t; do
@@ -43,7 +43,7 @@ for prec in double float int size_t; do
   * @brief MPI_Allreduce (MPI_MIN) for scalar. Combines values from all processes and distributes the result back to all processes.
   * @param val scalar value
   */ "
-  echo "[[nodiscard]] $prec Allreduce_min ($prec val) const;"
+  echo "[[nodiscard]] $prec Allreduce_min($prec val) const;"
 done
 
 ## Bcast
@@ -64,7 +64,7 @@ for prec in double float; do
   * @brief MPI_Bcast, Broadcasts a message from the process with rank "root" to all other processes
   * @param vec monolish vector (size N)
   * @param root root rank number
-  * @param gpu_sync sync gpu data.
+  * @param gpu_sync sync gpu data.This option does not work because std::vector is not support GPU.
   * @warning
   * When "only_cpu flag is enabled" and "send data is on the GPU", data is received from the GPU, 
   * MPI communication is performed, and finally data is sent to the GPU.
@@ -91,7 +91,7 @@ for prec in double float; do
   * @brief MPI_Gather, Gathers vector from all processes
   * The data is evenly divided and transmitted to each process.
   * @param sendvec send data, monolish vector (size N)
-  * @param recvvec recv data, std::vector (size N * # of procs)
+  * @param recvvec recv data, monolish vector (size N * # of procs)
   * @param val root rank number
   * @param gpu_sync sync gpu data. It receives sendvec, then performs MPI communication, and finally sends recvvec.
   * @warning
@@ -107,8 +107,8 @@ for prec in double float int size_t; do
   /**
   * @brief MPI_Gather, Gathers vector from all processes
   * The data is evenly divided and transmitted to each process.
-  * @param sendvec send data, monolish vector (size N)
-  * @param recvvec recv data, monolish vector (size N * # of procs)
+  * @param sendvec send data, std vector (size N)
+  * @param recvvec recv data, std vector (size N * # of procs)
   * @param val root rank number
   * @param gpu_sync sync gpu data. This option does not work because std::vector is not support GPU
   */