+++ include/mpi/monolish_mpi_core.hpp
@@ -92,6 +92,248 @@ public:
   void Barrier() const;
 
   /**
+   * @brief MPI_Send for scalar. Performs a blocking send.
+   * @param val scalar value
+   * @param dst rank of dstination
+   * @param tag message tag
+   * @param gpu_sync sync gpu data. This option does not work because scalar is
+   * automatically synchronized.
+   */
+  void Send(double val, int dst, int tag, bool only_cpu = false) const;
+
+  /**
+   * @brief MPI_Send for scalar. Performs a blocking send.
+   * @param val scalar value
+   * @param dst rank of dstination
+   * @param tag message tag
+   * @param gpu_sync sync gpu data. This option does not work because scalar is
+   * automatically synchronized.
+   */
+  void Send(float val, int dst, int tag, bool only_cpu = false) const;
+
+  /**
+   * @brief MPI_Send for scalar. Performs a blocking send.
+   * @param val scalar value
+   * @param dst rank of dstination
+   * @param tag message tag
+   * @param gpu_sync sync gpu data. This option does not work because scalar is
+   * automatically synchronized.
+   */
+  void Send(int val, int dst, int tag, bool only_cpu = false) const;
+
+  /**
+   * @brief MPI_Send for scalar. Performs a blocking send.
+   * @param val scalar value
+   * @param dst rank of dstination
+   * @param tag message tag
+   * @param gpu_sync sync gpu data. This option does not work because scalar is
+   * automatically synchronized.
+   */
+  void Send(size_t val, int dst, int tag, bool only_cpu = false) const;
+
+  /**
+   * @brief MPI_Send for scalar. Performs a blocking send.
+   * @param vec std::vector (size N)
+   * @param dst rank of dstination
+   * @param tag message tag
+   * @param gpu_sync sync gpu data.This option does not work because std::vector
+   * is not support GPU.
+   */
+  void Send(std::vector<double> &vec, int dst, int tag,
+            bool only_cpu = false) const;
+
+  /**
+   * @brief MPI_Send for scalar. Performs a blocking send.
+   * @param vec std::vector (size N)
+   * @param dst rank of dstination
+   * @param tag message tag
+   * @param gpu_sync sync gpu data.This option does not work because std::vector
+   * is not support GPU.
+   */
+  void Send(std::vector<float> &vec, int dst, int tag,
+            bool only_cpu = false) const;
+
+  /**
+   * @brief MPI_Send for scalar. Performs a blocking send.
+   * @param vec std::vector (size N)
+   * @param dst rank of dstination
+   * @param tag message tag
+   * @param gpu_sync sync gpu data.This option does not work because std::vector
+   * is not support GPU.
+   */
+  void Send(std::vector<int> &vec, int dst, int tag,
+            bool only_cpu = false) const;
+
+  /**
+   * @brief MPI_Send for scalar. Performs a blocking send.
+   * @param vec std::vector (size N)
+   * @param dst rank of dstination
+   * @param tag message tag
+   * @param gpu_sync sync gpu data.This option does not work because std::vector
+   * is not support GPU.
+   */
+  void Send(std::vector<size_t> &vec, int dst, int tag,
+            bool only_cpu = false) const;
+
+  /**
+   * @brief MPI_Send for scalar. Performs a blocking send.
+   * @param vec std::vector (size N)
+   * @param dst rank of dstination
+   * @param tag message tag
+   * @param gpu_sync sync gpu data. It receives sendvec, then performs MPI
+   * communication, and finally sends recvvec.
+   * @warning
+   * When only_cpu flag is enabled and send data is on the GPU, data is received
+   * from the GPU, MPI communication is performed, and finally data is sent to
+   * the GPU. If there is no GPU, or if there is no data on the GPU, no error
+   * will occur even if this flag is set to true.
+   */
+  void Send(monolish::vector<double> &vec, int dst, int tag,
+            bool only_cpu = false) const;
+
+  /**
+   * @brief MPI_Send for scalar. Performs a blocking send.
+   * @param vec std::vector (size N)
+   * @param dst rank of dstination
+   * @param tag message tag
+   * @param gpu_sync sync gpu data. It receives sendvec, then performs MPI
+   * communication, and finally sends recvvec.
+   * @warning
+   * When only_cpu flag is enabled and send data is on the GPU, data is received
+   * from the GPU, MPI communication is performed, and finally data is sent to
+   * the GPU. If there is no GPU, or if there is no data on the GPU, no error
+   * will occur even if this flag is set to true.
+   */
+  void Send(monolish::vector<float> &vec, int dst, int tag,
+            bool only_cpu = false) const;
+
+  /**
+   * @brief MPI_Recv for scalar. Performs a blocking recv.
+   * @param val scalar value
+   * @param src rank of source
+   * @param tag message tag
+   * @param gpu_sync sync gpu data. This option does not work because scalar is
+   * automatically synchronized.
+   * @return MPI status object
+   */
+  MPI_Status Recv(double val, int src, int tag, bool only_cpu = false) const;
+
+  /**
+   * @brief MPI_Recv for scalar. Performs a blocking recv.
+   * @param val scalar value
+   * @param src rank of source
+   * @param tag message tag
+   * @param gpu_sync sync gpu data. This option does not work because scalar is
+   * automatically synchronized.
+   * @return MPI status object
+   */
+  MPI_Status Recv(float val, int src, int tag, bool only_cpu = false) const;
+
+  /**
+   * @brief MPI_Recv for scalar. Performs a blocking recv.
+   * @param val scalar value
+   * @param src rank of source
+   * @param tag message tag
+   * @param gpu_sync sync gpu data. This option does not work because scalar is
+   * automatically synchronized.
+   * @return MPI status object
+   */
+  MPI_Status Recv(int val, int src, int tag, bool only_cpu = false) const;
+
+  /**
+   * @brief MPI_Recv for scalar. Performs a blocking recv.
+   * @param val scalar value
+   * @param src rank of source
+   * @param tag message tag
+   * @param gpu_sync sync gpu data. This option does not work because scalar is
+   * automatically synchronized.
+   * @return MPI status object
+   */
+  MPI_Status Recv(size_t val, int src, int tag, bool only_cpu = false) const;
+
+  /**
+   * @brief MPI_Recv for scalar. Performs a blocking recv.
+   * @param vec std::vector (size N)
+   * @param src rank of source
+   * @param tag message tag
+   * @param gpu_sync sync gpu data.This option does not work because std::vector
+   * is not support GPU.
+   * @return MPI status object
+   */
+  MPI_Status Recv(std::vector<double> &vec, int src, int tag,
+                  bool only_cpu = false) const;
+
+  /**
+   * @brief MPI_Recv for scalar. Performs a blocking recv.
+   * @param vec std::vector (size N)
+   * @param src rank of source
+   * @param tag message tag
+   * @param gpu_sync sync gpu data.This option does not work because std::vector
+   * is not support GPU.
+   * @return MPI status object
+   */
+  MPI_Status Recv(std::vector<float> &vec, int src, int tag,
+                  bool only_cpu = false) const;
+
+  /**
+   * @brief MPI_Recv for scalar. Performs a blocking recv.
+   * @param vec std::vector (size N)
+   * @param src rank of source
+   * @param tag message tag
+   * @param gpu_sync sync gpu data.This option does not work because std::vector
+   * is not support GPU.
+   * @return MPI status object
+   */
+  MPI_Status Recv(std::vector<int> &vec, int src, int tag,
+                  bool only_cpu = false) const;
+
+  /**
+   * @brief MPI_Recv for scalar. Performs a blocking recv.
+   * @param vec std::vector (size N)
+   * @param src rank of source
+   * @param tag message tag
+   * @param gpu_sync sync gpu data.This option does not work because std::vector
+   * is not support GPU.
+   * @return MPI status object
+   */
+  MPI_Status Recv(std::vector<size_t> &vec, int src, int tag,
+                  bool only_cpu = false) const;
+
+  /**
+   * @brief MPI_Recv for scalar. Performs a blocking recv.
+   * @param vec std::vector (size N)
+   * @param src rank of source
+   * @param tag message tag
+   * @param gpu_sync sync gpu data. It receives recvvec, then performs MPI
+   * communication, and finally recvs recvvec.
+   * @return MPI status object
+   * @warning
+   * When only_cpu flag is enabled and recv data is on the GPU, data is received
+   * from the GPU, MPI communication is performed, and finally data is sent to
+   * the GPU. If there is no GPU, or if there is no data on the GPU, no error
+   * will occur even if this flag is set to true.
+   */
+  MPI_Status Recv(monolish::vector<double> &vec, int src, int tag,
+                  bool only_cpu = false) const;
+
+  /**
+   * @brief MPI_Recv for scalar. Performs a blocking recv.
+   * @param vec std::vector (size N)
+   * @param src rank of source
+   * @param tag message tag
+   * @param gpu_sync sync gpu data. It receives recvvec, then performs MPI
+   * communication, and finally recvs recvvec.
+   * @return MPI status object
+   * @warning
+   * When only_cpu flag is enabled and recv data is on the GPU, data is received
+   * from the GPU, MPI communication is performed, and finally data is sent to
+   * the GPU. If there is no GPU, or if there is no data on the GPU, no error
+   * will occur even if this flag is set to true.
+   */
+  MPI_Status Recv(monolish::vector<float> &vec, int src, int tag,
+                  bool only_cpu = false) const;
+
+  /**
    * @brief MPI_Allreduce (MPI_SUM) for scalar. Combines values from all
    * processes and distributes the result back to all processes.
    * @param val scalar value
@@ -276,7 +518,8 @@ public:
    * all other processes
    * @param vec monolish vector (size N)
    * @param root root rank number
-   * @param gpu_sync sync gpu data.
+   * @param gpu_sync sync gpu data.This option does not work because std::vector
+   * is not support GPU.
    * @warning
    * When only_cpu flag is enabled and send data is on the GPU, data is received
    * from the GPU, MPI communication is performed, and finally data is sent to
@@ -291,7 +534,8 @@ public:
    * all other processes
    * @param vec monolish vector (size N)
    * @param root root rank number
-   * @param gpu_sync sync gpu data.
+   * @param gpu_sync sync gpu data.This option does not work because std::vector
+   * is not support GPU.
    * @warning
    * When only_cpu flag is enabled and send data is on the GPU, data is received
    * from the GPU, MPI communication is performed, and finally data is sent to
@@ -345,7 +589,7 @@ public:
    * @brief MPI_Gather, Gathers vector from all processes
    * The data is evenly divided and transmitted to each process.
    * @param sendvec send data, monolish vector (size N)
-   * @param recvvec recv data, std::vector (size N * # of procs)
+   * @param recvvec recv data, monolish vector (size N * # of procs)
    * @param val root rank number
    * @param gpu_sync sync gpu data. It receives sendvec, then performs MPI
    * communication, and finally sends recvvec.
@@ -363,7 +607,7 @@ public:
    * @brief MPI_Gather, Gathers vector from all processes
    * The data is evenly divided and transmitted to each process.
    * @param sendvec send data, monolish vector (size N)
-   * @param recvvec recv data, std::vector (size N * # of procs)
+   * @param recvvec recv data, monolish vector (size N * # of procs)
    * @param val root rank number
    * @param gpu_sync sync gpu data. It receives sendvec, then performs MPI
    * communication, and finally sends recvvec.
@@ -380,8 +624,8 @@ public:
   /**
    * @brief MPI_Gather, Gathers vector from all processes
    * The data is evenly divided and transmitted to each process.
-   * @param sendvec send data, monolish vector (size N)
-   * @param recvvec recv data, monolish vector (size N * # of procs)
+   * @param sendvec send data, std vector (size N)
+   * @param recvvec recv data, std vector (size N * # of procs)
    * @param val root rank number
    * @param gpu_sync sync gpu data. This option does not work because
    * std::vector is not support GPU
@@ -392,8 +636,8 @@ public:
   /**
    * @brief MPI_Gather, Gathers vector from all processes
    * The data is evenly divided and transmitted to each process.
-   * @param sendvec send data, monolish vector (size N)
-   * @param recvvec recv data, monolish vector (size N * # of procs)
+   * @param sendvec send data, std vector (size N)
+   * @param recvvec recv data, std vector (size N * # of procs)
    * @param val root rank number
    * @param gpu_sync sync gpu data. This option does not work because
    * std::vector is not support GPU
@@ -404,8 +648,8 @@ public:
   /**
    * @brief MPI_Gather, Gathers vector from all processes
    * The data is evenly divided and transmitted to each process.
-   * @param sendvec send data, monolish vector (size N)
-   * @param recvvec recv data, monolish vector (size N * # of procs)
+   * @param sendvec send data, std vector (size N)
+   * @param recvvec recv data, std vector (size N * # of procs)
    * @param val root rank number
    * @param gpu_sync sync gpu data. This option does not work because
    * std::vector is not support GPU
@@ -416,8 +660,8 @@ public:
   /**
    * @brief MPI_Gather, Gathers vector from all processes
    * The data is evenly divided and transmitted to each process.
-   * @param sendvec send data, monolish vector (size N)
-   * @param recvvec recv data, monolish vector (size N * # of procs)
+   * @param sendvec send data, std vector (size N)
+   * @param recvvec recv data, std vector (size N * # of procs)
    * @param val root rank number
    * @param gpu_sync sync gpu data. This option does not work because
    * std::vector is not support GPU