+++ include/mpi/mpi_dummy.hpp
@@ -2,9 +2,6 @@
 #ifndef OMPI_MPI_H
 #define OMPI_MPI_H
 
-typedef struct ompi_communicator_t *MPI_Comm;
-typedef struct ompi_op_t *MPI_Op;
-
 #ifndef OMPI_DECLSPEC
 #if defined(WIN32) || defined(_WIN32)
 #if defined(OMPI_IMPORTS)
@@ -36,6 +33,23 @@ OMPI_DECLSPEC extern struct ompi_predefined_op_t ompi_mpi_op_maxloc;
 OMPI_DECLSPEC extern struct ompi_predefined_op_t ompi_mpi_op_minloc;
 OMPI_DECLSPEC extern struct ompi_predefined_op_t ompi_mpi_op_replace;
 OMPI_DECLSPEC extern struct ompi_predefined_op_t ompi_mpi_op_no_op;
+
+using MPI_Status = int;
+// typedef struct ompi_status_public_t MPI_Status;
+typedef struct ompi_communicator_t *MPI_Comm;
+typedef struct ompi_datatype_t *MPI_Datatype;
+typedef struct ompi_errhandler_t *MPI_Errhandler;
+typedef struct ompi_file_t *MPI_File;
+typedef struct ompi_group_t *MPI_Group;
+typedef struct ompi_info_t *MPI_Info;
+typedef struct ompi_op_t *MPI_Op;
+typedef struct ompi_request_t *MPI_Request;
+typedef struct ompi_message_t *MPI_Message;
+typedef struct ompi_win_t *MPI_Win;
+typedef struct mca_base_var_enum_t *MPI_T_enum;
+typedef struct ompi_mpit_cvar_handle_t *MPI_T_cvar_handle;
+typedef struct mca_base_pvar_handle_t *MPI_T_pvar_handle;
+typedef struct mca_base_pvar_session_t *MPI_T_pvar_session;
 
 #if !OMPI_BUILDING
 #if defined(c_plusplus) || defined(__cplusplus)