+++ src/mpi/Bcast.hpp
@@ -38,7 +38,7 @@ void Bcast_core(monolish::vector<T> &vec, int root, MPI_Comm comm,
   logger.util_in(monolish_func);
 
 #if defined MONOLISH_USE_GPU
-  if (gpu_sync == true) {
+  if ((gpu_sync == true) && (vec.get_device_mem_stat() == true)) {
     vec.recv();
   }
 #endif